Small code cleanups.
git-svn-id: svn+ssh://svn.oneunified.net/var/local/svn/private/projects/oneunified/projects/trunk/TradeFrame@835 43143c16-083f-4785-a0a0-245a1fb241a9

+++ lib/OUBerkeleyDb/KeyValuePair.cpp
@@ -241,7 +241,7 @@ double CKeyValuePair::GetDouble(const std::string &key) {
 
 CKeyValuePair::structValue CKeyValuePair::GetBlob(const std::string &key) {
   structKey k( Blob, key.size(), key.c_str() );
-  u_int32_t size;
+  size_t size;
   void *p;
   Get( k, &p, &size );
   structValue v( p, size );
